DigitalOcean Docs / Rename DigitalOcean to dobs
This patch adds documentation for the DigitalOcean Block Storage driver
as well as renaming said driver to `dobs`.
