+++ drivers/storage/dobs/executor/dobs_executor.go
@@ -1,4 +1,4 @@
-// +build !libstorage_storage_executor libstorage_storage_executor_digitalocean
+// +build !libstorage_storage_executor libstorage_storage_executor_dobs
 
 package executor
 
@@ -11,8 +11,8 @@ import (
 	gofig "github.com/akutz/gofig/types"
 	"github.com/codedellemc/libstorage/api/registry"
 	"github.com/codedellemc/libstorage/api/types"
-	do "github.com/codedellemc/libstorage/drivers/storage/digitalocean"
-	doUtils "github.com/codedellemc/libstorage/drivers/storage/digitalocean/utils"
+	do "github.com/codedellemc/libstorage/drivers/storage/dobs"
+	doUtils "github.com/codedellemc/libstorage/drivers/storage/dobs/utils"
 )
 
 var (
