+++ drivers/storage/dobs/tests/terraform/templates/cloudinit.tpl
