+++ drivers/storage/dobs/utils/client.go
@@ -1,4 +1,4 @@
-// +build !libstorage_storage_driver libstorage_storage_driver_digitalocean
+// +build !libstorage_storage_driver libstorage_storage_driver_dobs
 
 package utils
