+++ drivers/storage/dobs/utils/utils_go1.7.go
@@ -1,5 +1,5 @@
 // +build go1.7
-// +build !libstorage_storage_driver libstorage_storage_driver_digitalocean
+// +build !libstorage_storage_driver libstorage_storage_driver_dobs
 
 package utils
