+++ imports/remote/imports_remote.go
@@ -5,7 +5,7 @@ package remote
 import (
 	// import to load
 	_ "github.com/codedellemc/libstorage/drivers/storage/azureud/storage"
-	_ "github.com/codedellemc/libstorage/drivers/storage/digitalocean/storage"
+	_ "github.com/codedellemc/libstorage/drivers/storage/dobs/storage"
 	_ "github.com/codedellemc/libstorage/drivers/storage/ebs/storage"
 	_ "github.com/codedellemc/libstorage/drivers/storage/efs/storage"
 	_ "github.com/codedellemc/libstorage/drivers/storage/fittedcloud/storage"
