+++ imports/remote/imports_remote_digitalocean.go
@@ -1,8 +1,8 @@
-// +build libstorage_storage_driver,libstorage_storage_driver_digitalocean
+// +build libstorage_storage_driver,libstorage_storage_driver_dobs
 
 package remote
 
 import (
 	// load the packages
-	_ "github.com/codedellemc/libstorage/drivers/storage/digitalocean/storage"
+	_ "github.com/codedellemc/libstorage/drivers/storage/dobs/storage"
 )
